requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   rq
    #   typer
    #   uvicorn
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi[all]==0.115.12
    # via -r requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.3.2
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via -r requirements.in
grpcio==1.71.0
    # via qdrant-client
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx[http2]==0.28.1
    # via
    #   fastapi
    #   ollama
    #   openai
    #   qdrant-client
huggingface-hub==0.30.2
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.6
    # via
    #   fastapi
    #   torch
jiter==0.9.0
    # via openai
joblib==1.5.0
    # via scikit-learn
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.5
    # via
    #   qdrant-client
    #   scikit-learn
    #   scipy
    #   transformers
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
ollama==0.4.8
    # via -r requirements.in
openai==1.77.0
    # via -r requirements.in
orjson==3.10.18
    # via
    #   -r requirements.in
    #   fastapi
packaging==25.0
    # via
    #   huggingface-hub
    #   transformers
pillow==11.2.1
    # via sentence-transformers
portalocker==2.10.1
    # via qdrant-client
protobuf==6.30.2
    # via qdrant-client
pydantic==2.11.4
    # via
    #   -r requirements.in
    #   fastapi
    #   ollama
    #   openai
    #   pydantic-extra-types
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.33.2
    # via pydantic
pydantic-extra-types==2.10.4
    # via fastapi
pydantic-settings==2.9.1
    # via fastapi
pygments==2.19.1
    # via rich
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   fastapi
    #   huggingface-hub
    #   transformers
    #   uvicorn
qdrant-client==1.14.2
    # via -r requirements.in
redis==6.0.0
    # via
    #   -r requirements.in
    #   rq
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.5
    # via fastapi-cli
rq==2.3.2
    # via -r requirements.in
safetensors==0.5.3
    # via transformers
scikit-learn==1.6.1
    # via sentence-transformers
scipy==1.15.2
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==4.1.0
    # via -r requirements.in
shellingham==1.5.4
    # via typer
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.1
    # via
    #   -r requirements.in
    #   transformers
torch==2.7.0
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.51.3
    # via
    #   -r requirements.in
    #   sentence-transformers
triton==3.3.0
    # via torch
typer==0.15.3
    # via fastapi-cli
typing-extensions==4.13.2
    # via
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   sentence-transformers
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
ujson==5.10.0
    # via fastapi
urllib3==2.4.0
    # via
    #   qdrant-client
    #   requests
uvicorn[standard]==0.34.2
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchdog==6.0.0
    # via -r requirements.in
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
